encrypt on user trusted device (#6)
* document and fix `encrypt` use case.

specifying an `output_file` is mandatory so that the plaintext is not overwritten. Once overwritten with a ciphertext, the data owner will not have the all the keys required to decrypt!

* use public parameters for encrypt.

public params are loaded from a base64-encoded file.

* remove flaky test.

test_ciphertext_is_never_equal_to_plaintext gets marked as flaky and does not add much value over the round trip test.